workable code

+++ data.py
@@ -7,7 +7,7 @@ output_size = 10
 
 def one_hot(data, depth):
     n = data.size
-    result = np.zeros([n, depth], dtype=float)
+    result = np.zeros([n, depth], dtype=np.float32)
     result[np.arange(n), data] = 1
     return result
